requirements.txt
PyYAML==5.4.1
rfc3986==1.5.0
schema==0.7.4
